requirements.txt
python-multipart
tensorflow-cpu
numpy
Pillow
requests
scikit-learn
seaborn
matplotlib